Add main source code
Add main analysis and diagnostic code.

For the record, most of this code was originally written and used about two years ago but not hitherto maintained in a repository. These versions have been cleaned up/better documented and will continue to be checked as data processing/plotting scripts are added in forthcoming commits.