Simplify CICE data loading module for regional diagnostics
Change the function get_processed_data_regional() in module src.data.cice to use the inputs directly for data directory names rather than modifying them in place under presumed directory structure, and to require netCDF variable name as an input parameter. In the same module, add an optional parameter for the region netCDF names in function get_region_masks(), falling back to the configuration values if not passed in, and fix a bug where 'nx' was written instead of 'ny'.