Change metadata for the region/sector for the GIN Seas
Change the sector/region labelled as 'gin_seas' in the region netCDF file to have long_name 'Greenland Sea' and short_name 'GRE' in module src/io/config.py. The actual key 'gin_seas' is not updated (e.g., to 'greenland_sea') because the region netCDF file has not been changed and the script to generate it is not currently in the repository.