Add script to calculate daily/monthly means of atmospheric forcing
* Remove calculation of daily and monthly means using CDO from interpolation script scripts/process_atmo_data/interp_to_cice_grid.sh
* Remove printing of output directory paths from python cleanup script scripts/process_atmo_data/_clean_interp.py
* Add new, separate python script to calculate daily and monthly means of atmospheric forcing fields which (1) correctly calculates the daily means for instantaneous fields and (2) saves all variables to one output file per year and per time averaging frequency
* Update function get_atmospheric_data_time_averages() in module src.data.atmo.py to match new file structure of daily/montly averages (specifically, to accept a list of variable names to load rather than one at a time).

Note that, for instantaneous fields, the prior calculation of daily means using CDO daymean calculates means of each set of 00, 03, 06, 09, 12, 15, 18, and 21 hour time steps, such that the 'daily mean' is effectively offset by -1.5 hours. The new python script calculates the daily mean centred on 12:00 of day N by linearly interpolating between consecutive 3 hour steps starting with the pair 00-03 of day N and ending with the pair 21 of day N to 00 of day N+1. For fields that are accumulations/averages, the daily averages are identical to the prior calculation with CDO.