Add script for detrended atmospheric fields and fix bug in VRILE diagnostics
* Add argument for netCDF data file name prefix to function get_atmospheric_data_time_averages() of module src/data/atmo.py, and remove automatic appending of '_d' or '_m' to netCDF variable names. Also add this argument to functions for computing diagnostics over VRILEs in module vrile_diagnostics.py.
* Fix bug in function compute_averages_over_vriles() in module vrile_diagnostics.py where the time normalisation flag is hitherto being interpreted in reverse such that True leads to no normalisation and vice-versa.
* Add script to calculate 2D detrended atmospheric forcing fields.
* Add atmospheric diagnostics, detrended and non-detrended versions, averaged over VRILEs to script scripts/vriles/vriles_cice.py, and clarify units of processed diagnostics averaged over VRILEs.

For the bug mentioned above, note that this does not affect prior results for classification of VRILEs since the the default parameter value (True) was implicitly being used in script scripts/vriles/vriles_cice.py, which was in fact not correct: the classification diagnostics should be integrated. Hence this should have been set to False, but in light of the above bug the classification was being correctly calculated anyway (i.e., two wrongs making a right in this case). The parameter is now explicitly passed in that script.